rename the test file for embeddings to follow the preferred structure
