+++ generative_ai/embeddings/multimodal_image_example.py
